feat: listAll method
